Allow multi-line string for category program list
descriptions,

Additionally, fix `display_menu` parsing to allow for these multi-line
strings.

Allow the `category_chooser` to take descriptions, to add some
little information such as (All-in-one) for some programs. It does this
by just taking the first word of the program entry description for the
lookup later on (to identify whether the program has been selected or
not). So the first word should always be just the program name, with the
exact same name as in the `programs` associative array entry.