MAJOR CHANGE: Rework category_chooser function
This rework allows the category chooser to accept any associative
arrays as exemplified in the `RNA_Questionnaire` function, instead of
only being set on variables from the rna-based programs and categories
as before.

This change is motivated by the fact that we need to run very similarly
a questionnaire for the DNA-based pipeline, and therefore not
duplicating the code would be ideal.

The `category_chooser` function instead takes 4 (associative) arrays as
function parameters, and uses these to do the same action as before.

The functions needed for this were also moved into the `utils.bash`
folder to ensure the `workflow.bash` script contains mainly
workflow-related code instead of helper functions.