A fully-working modularised pipeline
Each file contains its own workflow, specific to the sort of "general"
step that each one is responsible for i.e preprocessing, mapping or
assembly.

The current state also allows for the processing of multiple assembly
methods (reference-based or denovo) as well as choosing which methods to
use in the configuration file. The `main.nf` is only responsible for
calling these sub-workflow files. This simplifies the content by a large
amount.