fix: Bamsifter/htslib compilation errors
During the `docker build` command, it would consistently error out on
the recursive submodule building. This is because one submodule that was
loaded (`trinity-plugins/bamsifter`) itself had a script named
`build_htslib.sh` which would then build `htslib` in its own directory.

The modifications here result in changing the `bamsifter` submodule to
point to the `devel` branch, which instead of having `htslib` as a
submodule, it instead just packages a `.tar` file of htslib. This still
was not enough to build the entire package properly.

Additional measurements taken were to remove the commented-out
`autoconf`/`autoheader` lines which were needed to properly compile
htslib.

These steps then allowed for the proper compilation of the programs.