SplitNCigarReads parallelisation
The SplitNCigarReads process and the call-site was modified so that
instead of just working on the entirety and doing the processing on
more-or-less only a single-core, it is done per-chromosome. These can be
grouped in a list and the process is carried out for each grouping. This
is done via the `each` input directive. Afterwards, we group them by
sample name, sort them, mark the duplicates and index.